PACA-623: Switching to DSL generated entities for RDA API data.
