+++ apps/bfd-model/bfd-model-rda/mappings/FissProcCode.yaml
@@ -1,0 +1,60 @@
+mappings:
+- id: FissProcCode
+  messageClassName: gov.cms.mpsm.rda.v1.fiss.FissProcedureCode
+  entityClassName: gov.cms.bfd.model.rda.RdaFissProcCode
+  table:
+    name: fiss_proc_codes
+    schema: rda
+    primaryKeyColumns:
+    - dcn
+    - priority
+    columns:
+    - name: dcn
+      sqlType: varchar(23)
+      nullable: false
+      comment: 'Document Control Number: This field is the identification number for
+        a claim. It provides a reference number for the control and monitoring of
+        specific claims, and to differentiate between claims for the beneficiary;
+        it may be used to access data for the record depending on the copybook it
+        used for. This is a unique field created by the system.'
+    - name: priority
+      sqlType: smallint
+      javaType: short
+      nullable: false
+    - name: procCode
+      sqlType: varchar(10)
+      nullable: false
+      dbName: proc_code
+    - name: procFlag
+      sqlType: varchar(4)
+      dbName: proc_flag
+      comment: 'Procedure Flag: This field identifies a set of flags returned from
+        the Medicare Code Editor (MCE) module to indicate claim procedure code errors.'
+    - name: procDate
+      sqlType: date
+      dbName: proc_date
+    - name: rdaPosition
+      sqlType: smallint
+      javaType: short
+      nullable: false
+      dbName: rda_position
+    - name: lastUpdated
+      sqlType: timestamp with time zone
+      dbName: last_updated
+    comment: FISS Procedure Code
+  transformations:
+  - from: PARENT
+    to: dcn
+  - from: INDEX
+    to: priority
+  - from: procCd
+    to: procCode
+    optionalComponents: None
+  - from: procFlag
+  - from: procDt
+    to: procDate
+  - from: rdaPosition
+    transformer: UintToShort
+    optionalComponents: None
+  - from: NOW
+    to: lastUpdated
