+++ apps/bfd-model/bfd-model-rda/pom.xml
@@ -82,5 +82,25 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>gov.cms.bfd</groupId>
+                <artifactId>bfd-model-dsl-codegen-plugin</artifactId>
+                <version>${project.version}</version>
+                <configuration>
+                    <mappingPath>${project.basedir}/mappings</mappingPath>
+                    <outputDirectory>${project.build.directory}/generated-sources/entities</outputDirectory>
+                </configuration>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>entities</goal>
+                            <goal>sql</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
