--- apps/bfd-model/bfd-model-rda/src/main/java/gov/cms/bfd/model/rda/RdaMcsAudit.java
@@ -1,65 +1,0 @@
-package gov.cms.bfd.model.rda;
-
-import java.io.Serializable;
-import java.time.Instant;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.IdClass;
-import javax.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.FieldNameConstants;
-
-/** JPA class for the McsAudits table */
-@Entity
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@EqualsAndHashCode(onlyExplicitlyIncluded = true)
-@FieldNameConstants
-@IdClass(RdaMcsAudit.PK.class)
-@Table(name = "mcs_audits", schema = "rda")
-public class RdaMcsAudit {
-  @Id
-  @Column(name = "idr_clm_hd_icn", length = 15, nullable = false)
-  @EqualsAndHashCode.Include
-  private String idrClmHdIcn;
-
-  @Id
-  @Column(name = "priority", nullable = false)
-  @EqualsAndHashCode.Include
-  private short priority;
-
-  @Column(name = "rda_position")
-  private short rdaPosition;
-
-  @Column(name = "last_updated")
-  private Instant lastUpdated;
-
-  @Column(name = "idr_j_audit_num")
-  private Integer idrJAuditNum;
-
-  @Column(name = "idr_j_audit_ind", length = 1)
-  private String idrJAuditInd;
-
-  @Column(name = "idr_j_audit_disp", length = 1)
-  private String idrJAuditDisp;
-
-  /** PK class for the McsAudits table */
-  @Data
-  @NoArgsConstructor
-  @AllArgsConstructor
-  public static class PK implements Serializable {
-    private String idrClmHdIcn;
-
-    private short priority;
-  }
-}
