--- apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/test/java/gov/cms/bfd/pipeline/rda/grpc/source/DataTransformerTest.java
@@ -1,321 +1,0 @@
-package gov.cms.bfd.pipeline.rda.grpc.source;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.fail;
-
-import com.google.common.collect.ImmutableList;
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.OffsetDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.Consumer;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-/** Tests proper operation of the {@link DataTransformer} class. */
-public class DataTransformerTest {
-  /** The Transformer under test. */
-  private DataTransformer transformer;
-  /** The copied data to test. */
-  private List<Object> copied;
-
-  /** Sets the test up. */
-  @BeforeEach
-  public void setUp() {
-    transformer = new DataTransformer();
-    copied = new ArrayList<>();
-  }
-
-  /** Tests the {@link DataTransformer#validateAtLeastOneIsPresent} method. */
-  @Test
-  public void testValidateAtLeastOneIsPresent() {
-    assertEquals(
-        true, transformer.validateAtLeastOneIsPresent("ok-first-1", "1", "ok-second-1", ""));
-    assertEquals(
-        true, transformer.validateAtLeastOneIsPresent("ok-first-2", "", "ok-second-2", "2"));
-    assertEquals(
-        true, transformer.validateAtLeastOneIsPresent("ok-first-3", "1", "ok-second-3", null));
-    assertEquals(
-        true, transformer.validateAtLeastOneIsPresent("ok-first-4", null, "ok-second-4", "2"));
-    assertEquals(
-        true, transformer.validateAtLeastOneIsPresent("ok-first-5", "a", "ok-second-5", "b"));
-    assertEquals(
-        false,
-        transformer.validateAtLeastOneIsPresent("neither-first-1", "", "neither-second-1", ""));
-    assertEquals(
-        false,
-        transformer.validateAtLeastOneIsPresent("neither-first-2", null, "neither-second-2", ""));
-    assertEquals(
-        false,
-        transformer.validateAtLeastOneIsPresent("neither-first-3", "", "neither-second-3", null));
-    assertEquals(
-        false,
-        transformer.validateAtLeastOneIsPresent("neither-first-4", null, "neither-second-4", null));
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage(
-                "neither-first-1",
-                "expected either neither-first-1 or neither-second-1 to have value but neither did"),
-            new DataTransformer.ErrorMessage(
-                "neither-first-2",
-                "expected either neither-first-2 or neither-second-2 to have value but neither did"),
-            new DataTransformer.ErrorMessage(
-                "neither-first-3",
-                "expected either neither-first-3 or neither-second-3 to have value but neither did"),
-            new DataTransformer.ErrorMessage(
-                "neither-first-4",
-                "expected either neither-first-4 or neither-second-4 to have value but neither did")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyString} method. */
-  @Test
-  public void testCopyString() {
-    transformer
-        .copyString("length-one-ok", false, 1, 5, "1", copied::add)
-        .copyString("length-five-ok", false, 1, 5, "12345", copied::add)
-        .copyString("length-below-min", false, 2, 5, "1", copied::add)
-        .copyString("length-above-max", false, 1, 5, "123456", copied::add)
-        .copyString("null-ok", true, 1, 5, null, copied::add)
-        .copyString("null-bad", false, 1, 5, null, copied::add);
-
-    assertEquals(ImmutableList.of("1", "12345"), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage(
-                "length-below-min", "invalid length: expected=[2,5] actual=1"),
-            new DataTransformer.ErrorMessage(
-                "length-above-max", "invalid length: expected=[1,5] actual=6"),
-            new DataTransformer.ErrorMessage("null-bad", "is null")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyOptionalNonEmptyString} method. */
-  @Test
-  public void testCopyOptionalNonEmptyString() {
-    transformer
-        .copyOptionalNonEmptyString("value-not-present", 1, 5, () -> false, () -> null, copied::add)
-        .copyOptionalNonEmptyString("null-value-present", 1, 5, () -> true, () -> null, copied::add)
-        .copyOptionalNonEmptyString("empty-value-present", 1, 5, () -> true, () -> "", copied::add)
-        .copyOptionalNonEmptyString(
-            "non-empty-value-present", 1, 5, () -> true, () -> "A", copied::add)
-        .copyOptionalNonEmptyString("length-below-min", 2, 5, () -> true, () -> "1", copied::add)
-        .copyOptionalNonEmptyString(
-            "length-above-max", 2, 5, () -> true, () -> "123456", copied::add);
-
-    assertEquals(ImmutableList.of("A"), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage(
-                "length-below-min", "invalid length: expected=[2,5] actual=1"),
-            new DataTransformer.ErrorMessage(
-                "length-above-max", "invalid length: expected=[2,5] actual=6")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyCharacter} method. */
-  @Test
-  public void testCopyCharacter() {
-    transformer
-        .copyCharacter("length-one-ok", "1", copied::add)
-        .copyCharacter("length-below-min", "", copied::add)
-        .copyCharacter("length-above-max", "12", copied::add)
-        .copyCharacter("length-one-ok", "A", copied::add);
-    assertEquals(ImmutableList.of('1', 'A'), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage(
-                "length-below-min", "invalid length: expected=[1,1] actual=0"),
-            new DataTransformer.ErrorMessage(
-                "length-above-max", "invalid length: expected=[1,1] actual=2")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyDate} method. */
-  @Test
-  public void testCopyDate() {
-    transformer
-        .copyDate("valid-1", false, "2021-03-01", copied::add)
-        .copyDate("invalid-1", true, "2021/03/01", copied::add)
-        .copyDate("invalid-2", true, "2021-06-19T18:24:30", copied::add)
-        .copyDate("valid-2", true, "2021-10-21", copied::add)
-        .copyDate("null-ok", true, null, copied::add)
-        .copyDate("null-bad", false, null, copied::add);
-    assertEquals(ImmutableList.of(LocalDate.of(2021, 3, 1), LocalDate.of(2021, 10, 21)), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage("invalid-1", "invalid date"),
-            new DataTransformer.ErrorMessage("invalid-2", "invalid date"),
-            new DataTransformer.ErrorMessage("null-bad", "is null")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyTimestamp(String, boolean, String, Consumer)} method. */
-  @Test
-  public void testCopyTimestamp() {
-    final String VALID_ONE = "2021-03-01T01:01:01.1Z";
-    final String VALID_TWO = "2021-03-01T01:01:01.111111Z";
-    final String VALID_THREE = "2021-03-01T01:01:01.1+00:00";
-
-    transformer
-        .copyTimestamp("valid-1", false, VALID_ONE, copied::add)
-        .copyTimestamp("valid-2", true, VALID_TWO, copied::add)
-        .copyTimestamp("valid-3", true, VALID_THREE, copied::add)
-        .copyTimestamp("invalid-1", true, "20210301T010101Z", copied::add)
-        .copyTimestamp("invalid-2", true, "2021-03-01 01:01:01.1Z", copied::add)
-        .copyTimestamp("invalid-3", true, "2021-03-01T01:01:01.1", copied::add)
-        .copyTimestamp("null-ok", true, null, copied::add)
-        .copyTimestamp("null-bad", false, null, copied::add);
-
-    assertEquals(
-        List.of(
-            OffsetDateTime.parse(VALID_ONE).toInstant(),
-            OffsetDateTime.parse(VALID_TWO).toInstant(),
-            OffsetDateTime.parse(VALID_THREE).toInstant()),
-        copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage("invalid-1", "invalid timestamp"),
-            new DataTransformer.ErrorMessage("invalid-2", "invalid timestamp"),
-            new DataTransformer.ErrorMessage("invalid-3", "invalid timestamp"),
-            new DataTransformer.ErrorMessage("null-bad", "is null")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyAmount} method. */
-  @Test
-  public void testCopyAmount() {
-    transformer
-        .copyAmount("valid-1", false, "123.05", copied::add)
-        .copyAmount("invalid-1", true, "not a number", copied::add)
-        .copyAmount("invalid-2", true, "123a.00", copied::add)
-        .copyAmount("valid-2", true, "-456.98", copied::add)
-        .copyAmount("valid-3", true, "16", copied::add)
-        .copyAmount("null-ok", true, null, copied::add)
-        .copyAmount("null-bad", false, null, copied::add);
-    assertEquals(
-        ImmutableList.of(new BigDecimal("123.05"), new BigDecimal("-456.98"), new BigDecimal("16")),
-        copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage("invalid-1", "invalid amount"),
-            new DataTransformer.ErrorMessage("invalid-2", "invalid amount"),
-            new DataTransformer.ErrorMessage("null-bad", "is null")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyEnumAsCharacter} method. */
-  @Test
-  public void testCopyEnumAsCharacter() {
-    transformer
-        .copyEnumAsCharacter(
-            "no-value",
-            new EnumStringExtractor.Result(EnumStringExtractor.Status.NoValue),
-            copied::add)
-        .copyEnumAsCharacter(
-            "invalid-value",
-            new EnumStringExtractor.Result(EnumStringExtractor.Status.InvalidValue),
-            copied::add)
-        .copyEnumAsCharacter(
-            "null-value", new EnumStringExtractor.Result((String) null), copied::add)
-        .copyEnumAsCharacter("empty-string", new EnumStringExtractor.Result(""), copied::add)
-        .copyEnumAsCharacter("long-string", new EnumStringExtractor.Result("boo!"), copied::add)
-        .copyEnumAsCharacter("good-value", new EnumStringExtractor.Result("Z"), copied::add);
-    assertEquals(ImmutableList.of('Z'), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage("no-value", "no value set"),
-            new DataTransformer.ErrorMessage("invalid-value", "unrecognized enum value"),
-            new DataTransformer.ErrorMessage("null-value", "is null"),
-            new DataTransformer.ErrorMessage(
-                "empty-string", "invalid length: expected=[1,1] actual=0"),
-            new DataTransformer.ErrorMessage(
-                "long-string", "invalid length: expected=[1,1] actual=4")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyEnumAsString} method. */
-  @Test
-  public void testCopyEnumAsString() {
-    transformer
-        .copyEnumAsString(
-            "no-value",
-            false,
-            10,
-            new EnumStringExtractor.Result(EnumStringExtractor.Status.NoValue),
-            copied::add)
-        .copyEnumAsString(
-            "no-value-ok",
-            true,
-            10,
-            new EnumStringExtractor.Result(EnumStringExtractor.Status.NoValue),
-            copied::add)
-        .copyEnumAsString(
-            "invalid-value",
-            true,
-            10,
-            new EnumStringExtractor.Result(EnumStringExtractor.Status.InvalidValue),
-            copied::add)
-        .copyEnumAsString(
-            "null-value-ok", true, 10, new EnumStringExtractor.Result((String) null), copied::add)
-        .copyEnumAsString(
-            "null-value-bad", false, 10, new EnumStringExtractor.Result((String) null), copied::add)
-        .copyEnumAsString(
-            "unsupported-value-bad",
-            false,
-            10,
-            new EnumStringExtractor.Result(EnumStringExtractor.Status.UnsupportedValue, "boo!"),
-            copied::add)
-        .copyEnumAsString(
-            "good-value", false, 10, new EnumStringExtractor.Result("boo!"), copied::add);
-    assertEquals(ImmutableList.of("boo!"), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage("no-value", "no value set"),
-            new DataTransformer.ErrorMessage("invalid-value", "unrecognized enum value"),
-            new DataTransformer.ErrorMessage("null-value-bad", "is null"),
-            new DataTransformer.ErrorMessage("unsupported-value-bad", "unsupported enum value")),
-        transformer.getErrors());
-  }
-
-  /** Tests the {@link DataTransformer#copyStringWithExpectedValue} method. */
-  @Test
-  public void testCopyStringWithExpectedValue() {
-    transformer
-        .copyStringWithExpectedValue("both-null", true, 1, 1, null, null, copied::add)
-        .copyStringWithExpectedValue("both-same", true, 1, 10, "abcdef", "abcdef", copied::add)
-        .copyStringWithExpectedValue("too-short", true, 1, 10, "abcdef", "abc", copied::add)
-        .copyStringWithExpectedValue("too-long", true, 1, 10, "abcdef", "abcdefgh", copied::add)
-        .copyStringWithExpectedValue("start-mismatch", true, 1, 10, "abcdef", "qwcdef", copied::add)
-        .copyStringWithExpectedValue(
-            "middle-mismatch", true, 1, 10, "abcdef", "abqwef", copied::add)
-        .copyStringWithExpectedValue("end-mismatch", true, 1, 10, "abcdef", "abcdqw", copied::add)
-        .copyStringWithExpectedValue(
-            "long-random-mismatches",
-            true,
-            1,
-            30,
-            "abcdefghijklmnopqrstuvwxyz",
-            "a0cd4fg6ijklm4opqr5tuvw9yQ",
-            copied::add);
-    assertEquals(ImmutableList.of("abcdef"), copied);
-    assertEquals(
-        ImmutableList.of(
-            new DataTransformer.ErrorMessage("too-short", "value mismatch: masked=...---"),
-            new DataTransformer.ErrorMessage("too-long", "value mismatch: masked=......++"),
-            new DataTransformer.ErrorMessage("start-mismatch", "value mismatch: masked=##...."),
-            new DataTransformer.ErrorMessage("middle-mismatch", "value mismatch: masked=..##.."),
-            new DataTransformer.ErrorMessage("end-mismatch", "value mismatch: masked=....##"),
-            new DataTransformer.ErrorMessage(
-                "long-random-mismatches", "value mismatch: masked=.#..#..#.....#....#....#.#")),
-        transformer.getErrors());
-    try {
-      transformer.throwIfErrorsPresent();
-      fail("exception not thrown");
-    } catch (DataTransformer.TransformationException ex) {
-      assertEquals("failed with 6 errors", ex.getMessage());
-      assertEquals(ex.getErrors(), transformer.getErrors());
-    }
-  }
-}
