+++ apps/pom.xml
@@ -660,6 +660,8 @@ checkJavaFormat
 									<message>Undefined plugin versions can lead to un-reproducible
 										builds.</message>
 									<banTimestamps>false</banTimestamps>
+									<!-- Allow SNAPSHOT version of this plugin since we build it in our own project rather than pulling from a repository. -->
+									<unCheckedPluginList>gov.cms.bfd:bfd-model-dsl-codegen-plugin</unCheckedPluginList>
 								</requirePluginVersions>
 							</rules>
 						</configuration>
